Unbound key passthru: release before modifier
Got unbound key passthru while the keyboard is grabbed working fairly
well. It doesn't work well enough to type, but for using window manager
and application keybindings it's great.

Because pressing a key in xorg implicitly creates an active grab so that
the application receiving the KeyPress also receives the KeyRelease, we
can't simulate a keypress until the key is released. I have no reason to
think there's some way to disable this particular behaviour of xorg. The
main problem with sending keys with release is that people often release
the desired modifier keys after hitting the desired key but before
releasing it. It's the worst for me when I'm typing sentences quickly
and capitalizing some words. So ptrkeys isn't currently suitable for an
ancillary use case I imagined where the numpad was mapped to pointer
movement and the keyboard was permanently grabbed. The next step is to
remember the modifiers' state on KeyPress and try to recreate it on
KeyRelease---xdotool does it, and has only had a few bugs related to it.

I turned off key repeat again, but if remembering/recreating the
modifier state works I'll be able to leave it alone again.

various:

- Support modmap changes by finding numlockmask in the modmap updating
  the saved modifier keycodes on MappingNotify events.
- Disable passthru and warn the user if xtest isn't available.
- Add `Key.grab` so that umodified keys can be grabbed on start if desired
- Add `-std=c99 -pedantic` CFLAGS.
- Add `#define _XOPEN_SOURCE 700` to enable POSIX extensions,
  especially `clock_gettime`
- Remove some scroll bindings and constants. Scrolling using Button4 and
  Button5 worked in zathura or ghostview, I think, but not in firefox or
  chrome, so I'm going to add scrolling using XInput2 instead.
- One key can no longer have multiple bindings, due to how confusing it
  could be. Need to add a check for this.
- Add a provisional makefile.

dbg.h:

My modified version of Zed Shaw's dbg.h didn't compile with `-std=c99
-pedantic` because the macros were designed to sometimes have zero
variadic args, prompting a rework. I changed dbg.h enough that it
deserved a new name; I didn't want to have macros named log and logf,
due to functions with the same names in math.h, so I went with jot.

While it might occasionally be useful having errno in every log message
when it's nonzero, it isn't conventional to clear errno on success, so
library functions that succeed will sometimes leave a dirty errno.
Printing out irrelevant errno strings is confusing and noisy. As with
most things in C, we'll just have to be disciplined and make sure to
include errno after calling functions that return errors with it. Like
so:

    if (clock_gettime(CLOCK_MONOTONIC, &ts)) {
        jot("get time: %s", strerror(errno));
    }