Key event passthru experiments
After about a week of fussing with passthru I don't think it's worth it.
Some programs ignore XSendEvent, making it useless for this, and when using
xtest and ungrabbing/grabbing the keyboard it's too easy to miss a
KeyRelease event, corrupting your state.

All of my ideas of how to do it are hacks, anyhow, and ptrkeys will be
simpler without it.

                                  ***

I tried saving which key was down and adding a check to the event loop
to passthru a press/release pair of events at the server's repeat
interval. But after passing through any of the repeat events I no longer
get the release event corresponding to the physical release.

                                  ***

At first I thought I'd have to press/release modifier keys determined by
the mod map to passthru the modifier state, but ungrabbing the keyboard
doesn't change the keymap/modifier state.

                                  ***

I thought that if I only passed through a press event that
the xserver would somehow repeat the key for me "in the background"
until I sent a release event, but that's not the case. Since ptrkeys has
the keyboard grabbed, it gets the repeat events.